refactor(project): refactor all project to handle modules
