+++ packages/api/babel.config.js
@@ -1,0 +1,3 @@
+const config = require('@fullstack-playground/babel');
+
+module.exports = config;
