+++ packages/api/src/auth/getToken.ts
