+++ packages/api/src/database.ts
@@ -1,0 +1,19 @@
+import mongoose from 'mongoose';
+
+import { config } from './config';
+
+export const connectDatabase = () => {
+  return new Promise((resolve, reject) => {
+    mongoose.Promise = global.Promise;
+    mongoose.connection
+      .on('error', error => reject(error))
+      // eslint-disable-next-line
+      .on('close', () => console.log('Database connection closed.'))
+      .once('open', () => resolve(mongoose.connections[0]));
+
+    mongoose.connect(config.MONGO_URI, {
+      useNewUrlParser: true,
+      useCreateIndex: true,
+    });
+  });
+};
