+++ packages/api/src/models/index.ts
@@ -1,5 +1,2 @@
-import UserModel, { IUser } from '../modules/user/UserModel';
 import TenantModel, { ITenant } from '../modules/tenant/TenantModel';
-
-export { UserModel as User, IUser };
 export { TenantModel as Tenant, ITenant };
