+++ packages/babel/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@rest-api/babel",
+  "name": "@fullstack-playground/babel",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
@@ -18,6 +18,9 @@
     "@babel/preset-react": "7.7.4",
     "@babel/preset-typescript": "7.7.7",
     "babel-plugin-require-context-hook": "1.0.0"
+  },
+  "devDependencies": {
+    "@types/babel__core": "7.1.3"
   },
   "main": "index.js"
 }
