+++ packages/modules/README.md
@@ -1,0 +1,3 @@
+# @fullstack-playground/modules
+
+All fullstack playground models/modules to be consumed by Api, GraphQL and so on
