+++ packages/modules/src/index.ts
@@ -1,0 +1,1 @@
+export { default as UserModel, IUser } from './user/UserModel';
