--- packages/server/babel.config.js
@@ -1,3 +1,0 @@
-const config = require('@rest-api/babel') 
-
-module.exports = config;
