+++ packages/test/README.md
@@ -1,3 +1,3 @@
-# @rest-api/test
+# @fullstack-playground/test
 
 Test utilities to make testing easier
